신용대출상품 페이지 - 필터링 기능 구현 완료

+++ src/main/java/smu/toyproject1/entity/CreditLoanProduct.java
@@ -13,9 +13,9 @@ public class CreditLoanProduct {
     private String loanType;
     private String cbCompany;
     private String rateType;
-    private int averageRate;
+    private double averageRate;
 
-    public CreditLoanProduct(String company, String productName, String method, String loanType, String cbCompany, String rateType, int averageRate) {
+    public CreditLoanProduct(String company, String productName, String method, String loanType, String cbCompany, String rateType, double averageRate) {
         this.company = company;
         this.productName = productName;
         this.method = method;
